requirements.txt
appdirs==1.4.4
aiofiles==0.6.0
aiocqhttp==1.4.2
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
baike==1.5.0
beautifulsoup4==4.9.3
bilibili-api==5.1.1
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cssutils==2.2.0
cycler==0.10.0
docutils==0.17.1
fastapi==0.63.0
feedparser==6.0.2
gino==1.0.1
greenlet==1.0.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
ImageHash==4.2.0
importlib-metadata==4.0.1
jieba==0.42.1
keyring==23.0.1
kiwisolver==1.3.1
loguru==0.5.3
lxml==4.6.3
matplotlib==3.4.1
multidict==5.1.0
nest-asyncio==1.5.1
nonebot-adapter-cqhttp==2.0.0a12
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0a13.post1
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.3
Pillow==8.2.0
pkginfo==1.7.0
psutil==5.8.0
pydantic==1.8.1
pyee==8.1.0
Pygments==2.9.0
pygtrie==2.4.2
pyparsing==2.4.7
pypinyin==0.41.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
python-Levenshtein==0.12.2
PyYAML==5.4.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
scapy==2.4.5
scipy==1.6.2
sgmllib3k==1.0.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.3.24
starlette==0.13.6
tomlkit==0.7.0
tqdm==4.60.0
twine==3.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
webencodings==0.5.1
websockets==8.1
win32-setctime==1.0.3
yarl==1.6.3
zipp==3.4.1
ruamel.yaml==0.17.16
imageio==2.9.0
fuzzywuzzy==0.18.0